슬랙 정보 관리 방식 변경 - 테스트 메시지 전송 API 추가, 트랜잭션에서도 createDate 확인 가능하도록 수정

+++ src/main/java/com/example/weaponMaster/modules/slack/repository/UserSlackNoticeRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 public interface UserSlackNoticeRepository extends JpaRepository<UserSlackNotice, Integer> {
 
     @Query(value = "SELECT * FROM user_slack_notice WHERE user_id = :userId AND notice_type = :noticeType", nativeQuery = true)
-    UserSlackNotice findByUserIdAndType(String userId, Integer noticeType);
+    UserSlackNotice findByUserIdAndType(String userId, Byte noticeType);
 }
